♻️ Refatorando
