+++ frontend/.gitignore
@@ -33,3 +33,5 @@ yarn-error.*
 
 # typescript
 *.tsbuildinfo
+
+src/variables.ts
