+++ frontend/src/@types/culturaDto.ts
@@ -1,23 +1,37 @@
-export interface CulturaDto {
-    _id: string,
+interface  PontoCultivo {
     latitude: string;
     longitude: string;
+  }
+  
+  interface Temperatura {
+    data: string;
+    temperatura: number;
+  }
+  
+  interface Pluviometria {
+    data: string;
+    pluviometria: number;
+  }
+  
+  interface Alerta {
+    [data: string]: number;
+  }
+  
+  export interface Cultivo {
+    _id ?: string 
+    ponto_cultivo: PontoCultivo;
     nome_cultivo: string;
-    tempFrequency: string;
-    maxTemp: string;
-    minTemp: string;
-    pluviFrequency: string;
-    maxPluvi: string;
-    minPluvi: string;
-    temperaturas: Temperatura[],
-    pluviometrias: Pluviometria[],
-} 
+    temperatura_max: number;
+    pluviometria_max: number;
+    temperatura_min: number;
+    pluviometria_min: number;
+    temperaturas: Temperatura[];
+    pluviometrias: Pluviometria[];
+    alertasTemp: Alerta[];
+    alertasPluvi: Alerta[];
+    lastUpdate: string;
+  }
+  
 
-export interface Temperatura{
-    data: Date,
-    temperatura: number,
-}
-export interface Pluviometria{
-    data: Date,
-    pluviometria: number,
-}
+
+
